refactor: modify import statements and export function name
